Set scrape timeout to `50s` and add garden cluster ID as external label
The default value is 10s.
https://prometheus.io/docs/prometheus/latest/configuration/configuration/#configuration-file

From experience, we observed that the federation from the garden-prometheus to the
availability-prometheus (introduced in a future PR) fails, because the request duration can reach
the 10s, due to the metering metrics that are accumulated over the course
of a month for the deleted shoots. This can be especially worth in landscapes where
there are many short lived shoots.

By increasing the scrape_timeout, we can improve the reliability of
the federation from the garden-prometheus to the availability-prometheus.

A value is chosen that is slightly less than the scrape_interval (`1m`) to avoid
issues when consecutive scrapes are overlapping.