Scrape config for `{gardener,kube}-apiserver`
The virtual garden system component now contains also RBAC resources for
the garden Prometheus to be allowed to scrape the API servers.
The RBAC resources were put there to not get into the business of
creating a new `ManagedResource` just for the garden Prometheus that
would be deployed into the virtual garden cluster.

When the shoot Prometheus is migrated, we might move this out of there
again.