Scrape config for garden and aggregate Prometheus
The garden Prometheus scrapes all aggregate Prometheis via their Ingress exposure.

For the integration test, we have to add a fake CRD for the Seed API
because the garden controller will try to list seeds as part of its
reconciliation, but gardener-apiserver is not really running in the
integration test.