Redirect all non-vpn traffic on port 8132 of istio ingress gateway to https port.
Previously, port 8132 of the istio ingress gateway acted as an ordinary http server
answering queries for all resources with 404 resource not found. To make probing
more expensive and less convenient for potential attackers, this change now redirects
requests to the https port. The effect is that potential attackers need to invest
more cpu (for TLS) and do not get a proper response unless they set a correct SNI
header.