hack/kind-up: configure default StorageClass for kind cluster to use `local` volume
By default the rancher.io/local-path provisioner [1] uses `hostPath`, which does
not implement the Kubernetes volume.MetricsProvider interface.

Also, kind does not expose an easy way to configure the StorageClass, since it
uses a manifest, which is embedded in the node image [2].

This change configures the default StorageClass used by kind by annotating it
with `defaultVolumeType: local`, which implements the volume.MetricsProvider
interface, and would allow us to scrape the kubelets for the various
kubelet_volume_stats_* metrics.

[1]: https://github.com/rancher/local-path-provisioner
[2]: https://github.com/kubernetes-sigs/kind/blob/main/pkg/cluster/internal/create/actions/installstorage/storage.go