Remove the additional matcher from the Alertmanager config (#9384)
The default `AlertmanagerConfigMatcherStrategy` is `OnNamespace` and hence the
prometheus operator injects a label matcher matching the namespace of the
`AlertmanagerConfig` object for all its routes and inhibition rules.

This is not desired in our use cases (shoot, seed, garden alertmanagers).

This configuration option should be set to `None`: then the operator will not
add any additional matchers other than the ones specified in the
`AlertmanagerConfig`.

https://prometheus-operator.dev/docs/operator/api/#monitoring.coreos.com/v1.AlertmanagerConfigMatcherStrategy