Fix the match expression in the alertmanager configuration (#9387)
The type of `AlertmanagerConfigSpec.Route.Routes` is intended to be a self
referential type, ie. the type of `AlertmanagerConfigSpec.Route` itself.

Based on the comment here:

https://github.com/prometheus-operator/prometheus-operator/blob/07e4267d65e59adf503a352a08c6faead527540d/pkg/apis/monitoring/v1beta1/alertmanager_config_types.go#L121-L129

there is a limitation in the CRD schema such that this type constraint can not
be expressed and the generic JSON type is used instead for
`AlertmanagerConfigSpec.Route.Routes`.

This means that the compiler can not help with checking the field names of
`AlertmanagerConfigSpec.Route.Routes`.

It turns out that the old and deprecated Alertmanager configuration `match_re`
that was migrated does not exist on the `AlertmanagerConfigSpec`
abstraction level. It is expected because it is a deprecated field and
`matchers` should be used instead.

https://prometheus.io/docs/alerting/latest/configuration/#route

In the current implementation, the superfluous `match_re` field is not
considered silently, so the resulting route does not have any match filter
expressions. This means that although this is not the intended behavior, the
impact was more forwarded alerts (which is better) and not fewer or no forwarded
alerts at all.

This commit uses the `matchers` field with the name, matchType and value fields
instead.