local-path provisioner with support for volume resizing (#9363)
* hack/kind-up: configure default StorageClass for kind cluster to use `local` volume

By default the rancher.io/local-path provisioner [1] uses `hostPath`, which does
not implement the Kubernetes volume.MetricsProvider interface.

Also, kind does not expose an easy way to configure the StorageClass, since it
uses a manifest, which is embedded in the node image [2].

This change configures the default StorageClass used by kind by annotating it
with `defaultVolumeType: local`, which implements the volume.MetricsProvider
interface, and would allow us to scrape the kubelets for the various
kubelet_volume_stats_* metrics.

[1]: https://github.com/rancher/local-path-provisioner
[2]: https://github.com/kubernetes-sigs/kind/blob/main/pkg/cluster/internal/create/actions/installstorage/storage.go

* provider-local: use `volumeType: local` for the default StorageClass

* Set allowVolumeExpansion=true for default StorageClass on local setup

* hack/kind-up: add support for configuring local path provisioner with resize support

This commit adds a new option to hack/kind-up.sh which allows configuring the
local-path provisioner to be deployed with volume resize support.

* Delete the storageclass which comes with stock local-path provisioner

Delete the storageclass, which comes with the stock rancher/local-path
provisioner, when we are deploying the local cluster with volume resize support.

* Makefile: add support for deploying a local setup with volume resize support

In order to deploy a local dev setup with local-path provisioner with support
for resizing PVCs you should run:

    DEV_SETUP_WITH_LPP_RESIZE_SUPPORT make kind-up

* hack/kind-up: pass `--ignore-not-found` to `kubectl delete`

Do not fail if the default storageclass, which comes with rancher.io/local-path
provisioner does not exist when trying to delete it.

* provider-local: configure storageclass to be consistent with the kind cluster

Configure the StorageClass in `provider-local` to be consistent with the one
that we configure on the nodes in the kind cluster.

* tools: add `kustomize` dependency on `kind-up` target

* hack/tools: use `SYSTEM_NAME` and `SYSTEM_ARCH` where applicable

Also, add `kubectl` dependency to `kind-operator-up`

* Address PR feedback

- Fix indentation
- Remove leading underscores from function names
- Use `kubectl annotate` and `kubectl patch` where applicable

* hack/kind-up: use a custom image of the LPP with resize support

Instead of building it locally, we will use an already built one.