Fix kube-apiserver service exposure after replacement of ingress resources with direct istio exposure. (#9393)
The DNS wildcard entry for ingress domains pointing to nginx in the past now points to
the default istio ingress gateway. Therefore, all ingress resources need to be exposed
using the default istio ingress gateway. For single-zonal shoot clusters, this was no
longer the case after the ingress resource replacement with istio resources.
This caused the dashboard terminal to fail for those cluster, but succeed for highly-
available shoot clusters.
This change fixes this issue by switching to the default istio ingress gateway for
ingress domains.