Fix race condition in istio zone migration. (#9460)
The istio zone migration offers the possibility to migrate the shoot
exposure to a consistent istio namespace pattern without causing
disruption outside of the maintenance window for shoot cluster owners.
Unfortunately, the cleanup of orphaned istio resources may interfere
with the migration targets as it may detect the istio namespaces to
refer to a none-existing zone and be empty. Therefore, the namespace
may be cleaned up. As this happens directly before the creation in the
seed reconcile flow the seed cluster may end up with temporarily having
a deleted migration target namespace, which will be recreated again in
the next seed reconcile flow.
This change safe-guards the migration by ensuring that the migration
targets will be kept alive even if they do not see active usage.

Co-authored-by: Johannes Scheerer <johannes.scheerer@sap.com>