label workers with their expected operating system config key
Nodes with an outdated operating system config key are skipped while
checking whether a shoot is healthy. With the addition of kubeReserved
in the config key calculation, the kubeReserved value would have to be
annotated on the node objects to allow for this check. However, as this
leaks implementation details, just annotate each node with the expected
config key. This allows for future extensions without having to
introduce additional labels on the node objects.