[GEP-19] Integrate Prometheus and `blackbox-exporter` deployments into `Garden` controller (#9543)
* Set `spec.externalURL` when Prometheus is exposed via `Ingress`

This is to align the configuration with existing Prometheis

* Support Prometheus deployments with `replicas > 1`

- deploy PDB

GRM's HA webhook will make sure that proper TSCs get injected into the pods

* Allow overwriting `maxAllowed` for VPA

* Allow specifying CA name for signing self-signed server cert

* Make Prometheus component instantiation reusable

* Allow setting ingress fields after component has been initialized

Similar to https://github.com/gardener/gardener/blob/cae76199de1f652b5f27e601d7d9fb5b9058a743/pkg/component/observability/monitoring/alertmanager/component.go#L40-L47
Will also be needed later when `gardenlet` deploys the shoot Prometheus via this component.

* Integrate Prometheus deployment into `Garden` controller

* Set scrape timeout to `50s` and add garden cluster ID as external label

The default value is 10s.
https://prometheus.io/docs/prometheus/latest/configuration/configuration/#configuration-file

From experience, we observed that the federation from the garden-prometheus to the
availability-prometheus (introduced in a future PR) fails, because the request duration can reach
the 10s, due to the metering metrics that are accumulated over the course
of a month for the deleted shoots. This can be especially worth in landscapes where
there are many short lived shoots.

By increasing the scrape_timeout, we can improve the reliability of
the federation from the garden-prometheus to the availability-prometheus.

A value is chosen that is slightly less than the scrape_interval (`1m`) to avoid
issues when consecutive scrapes are overlapping.

* Adapt Plutono config to allow connection to garden Prometheus

* ------------ empty separator commit -------------

* Scrape config for etcd and backup-restore sidecar

The `monitoring/component.go` will soon disappear when `prometheus-operator` takes over management of the shoot Prometheus. Currently, this file would introduce an import cycle:

```
package github.com/gardener/gardener/cmd/gardener-operator
	imports github.com/gardener/gardener/cmd/gardener-operator/app
	imports github.com/gardener/gardener/pkg/operator/controller
	imports github.com/gardener/gardener/pkg/component/shared
	imports github.com/gardener/gardener/pkg/component/etcd/etcd
	imports github.com/gardener/gardener/pkg/component/observability/monitoring/prometheus/garden
	imports github.com/gardener/gardener/pkg/component/observability/monitoring/alertmanager
	imports github.com/gardener/gardener/pkg/component/observability/monitoring
	imports github.com/gardener/gardener/pkg/component/etcd/etcd: import cycle not allowed
```

Hence, let's remove the dependency to the `etcd` package to be able to move forward.

* Scrape config for `{gardener,kube}-apiserver`

The virtual garden system component now contains also RBAC resources for
the garden Prometheus to be allowed to scrape the API servers.
The RBAC resources were put there to not get into the business of
creating a new `ManagedResource` just for the garden Prometheus that
would be deployed into the virtual garden cluster.

When the shoot Prometheus is migrated, we might move this out of there
again.

* Scrape config for virtual cluster control plane

- GAC
- GCM
- GSCHED
- KCM

* Add a few more scrape configs

* Scrape config for garden and aggregate Prometheus

The garden Prometheus scrapes all aggregate Prometheis via their Ingress exposure.

For the integration test, we have to add a fake CRD for the Seed API
because the garden controller will try to list seeds as part of its
reconciliation, but gardener-apiserver is not really running in the
integration test.

* Add `PrometheusRule`s for garden Prometheus

* ------------ empty separator commit -------------

* Make `blackboxexporter` component reusable for seed cluster type

In this PR, we will deploy BBX into the garden cluster (similar cluster
type), but in a future PR, the shoot controller will also need to deploy
BBX into the seed cluster.

* Integrate `blackbox-exporter` deployment into `Garden` controller

* Add `blackbox-exporter` scrape configs for Garden scenario

* ------------ empty separator commit -------------

* Introduce `to-all-garden-scrape-targets` network policy label

* Address PR review feedback

* Delete global monitoring secret replica on garden deletion

Otherwise, it will remain in the system even after `Garden` got deleted,
causing issues with future `Garden` creations, see for example:
https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/9543/pull-gardener-e2e-kind-operator/1776264086145732608

* Address PR review feedback

* Address PR review feedback

* Address PR review feedback