Enable basic auth for Plutono + block admin API in Ingress
When the sidecar updates the dashboards in the file system, it needs to call an API of Plutono to make it reload the dashboards. This is only available in the admin API, and we need credentials to call it.
This was disabled previously. Now, we enable it and generate a basic auth secret which is auto-rotated each `30d`. The API is only accessible within the cluster itself, and blocked via Ingress. Also, no other pod can talk to Plutono (protected via network policies), hence, access to this API is basically only possible from within the pod (by the sidecar) and from human operators with port-forwarding.