Add label to dashboard `ConfigMap`
The dashboard `ConfigMap` created by Gardener will be picked up by the sidecar and no longer directly mounted into the pod (next commit).
For a few releases, when computing this `ConfigMap`, the old (now deprecated) extensions contract for providing dashboards will be respected, i.e., they will be collected and added to this `ConfigMap`.
When an extension switches to the approach, this will no longer happen, but the sidecar will pick it up.