[GEP-19] Add sidecar to Plutono for fetching dashboard `ConfigMap`s dynamically (#9624)
* Nit: Import alias + lower case variable

* Drop dashboard configmap separation for garden clusters

Previously, there were two "folders" for dashboards, "garden" and "global".
We don't follow this pattern for the seed or shoot plutono, so let's also drop it for the garden plutono for simplicity.
Now, all dashboards are added in the same folder.

* Minors nits/cosmetics/introduction of some constants

* Documentation

* Add dashboard refresher image

See https://github.com/kiwigrid/k8s-sidecar?tab=readme-ov-file#what for details how this works/what this does.

* Enable basic auth for Plutono + block admin API in Ingress

When the sidecar updates the dashboards in the file system, it needs to call an API of Plutono to make it reload the dashboards. This is only available in the admin API, and we need credentials to call it.
This was disabled previously. Now, we enable it and generate a basic auth secret which is auto-rotated each `30d`. The API is only accessible within the cluster itself, and blocked via Ingress. Also, no other pod can talk to Plutono (protected via network policies), hence, access to this API is basically only possible from within the pod (by the sidecar) and from human operators with port-forwarding.

* Add `ServiceAccount`, RBAC + access to runtime API server

The sidecar is a controller that watches `configmaps`, so it needs access/permissions to do so.

* Add label to dashboard `ConfigMap`

The dashboard `ConfigMap` created by Gardener will be picked up by the sidecar and no longer directly mounted into the pod (next commit).
For a few releases, when computing this `ConfigMap`, the old (now deprecated) extensions contract for providing dashboards will be respected, i.e., they will be collected and added to this `ConfigMap`.
When an extension switches to the approach, this will no longer happen, but the sidecar will pick it up.

* Add sidecar

* Adapt unit tests

* Address PR review feedback