Remote write configuration
Note: Support for providing Queue configuration via gardenlet config has been dropped because it was of type `*string` which is no longer working now with Golang structs. We would have to reference the `monitoringv1.QueueConfig` type in the gardenlet config, however, I'd like to prevent this for now. It's anyways unclear whether this is used at all, and if, we can discuss how to cleanly allow modifying the configuration here. A better approach (now that we have a `Prometheus` CRD) could be that an extension mutates it and injects whatever configuration it wants.
In that sense, we might want to entirely drop support for providing remote write config via gardenlet config, and move everything out to extension responsibilities.