Enable access log in istio ingress gateway for apiserver-proxy connections (#9686)
* Enable access log in istio ingress gateway for apiserver-proxy
connections.

Unfortunately, the connections initiated by apiserver-proxy did not make
it into the access log with the last access log changes. However, it
works as expected when the access log is directly configured for the
tcp proxy filter entry. Unfortunately, the log format needs to be
configured as well, but this allows us to display the internet routable
IP address instead of the source IP from the proxy protocol.

* Address review feedback.

Add `apiserver-proxy` suffix to the access log line.

Co-authored-by: Axel Siebenborn <axel.siebenborn@sap.com>

---------

Co-authored-by: Axel Siebenborn <axel.siebenborn@sap.com>