Make sure `max_over_time` is used in the alerts in case of flapping
Without it, flapping conditions would probably go undetected because
the alerts would be reset each time.

For example, when the seed is reconciling its artifacts, some of its
managed resources can enter the progressing state for a brief period,
and to avoid alerts from being restarted, the `max_over_time` function
is now introduced to smooth out this behaviour. This is combined with
the subquery syntax [1] to query over the last minutes only.

This change is also reducing the time before firing an alert in the `for`
entry from 30m to 10m for some alerts to alert more aggressively if any of
those conditions fail. An exception is the alert `SeedAPIServerUnavailable`,
which was already firing after 2m, but now changed to 3m to preserve the
same behaviour after the change introducing `max_over_time` to smooth
out possible flappings.

[1] https://prometheus.io/docs/prometheus/latest/querying/basics/#subquery