WIP: Option 2: use the host name to scrape the nodes
The Prometheus federation fails with the following error
in the local setup:

      Get "https://172.18.0.2:10250/metrics":
      tls: failed to verify certificate:
      x509: cannot validate certificate for 172.18.0.2
      because it doesn't contain any IP SANs

We could tweak the Prometheus configuration to scrape the nodes
using the hostname and not the node's IP. That could be acceptable
(although not idiomatic) in real clusters as well.

This way Prometheus can validate the certificate, because it connects
via the hostname, and the certificate contains a DNS SAN entry
both in the local setup and in real clusters.

Local setup: only DNS

  openssl s_client -connect 172.18.0.3:10250 -showcerts </dev/null 2>/dev/null | openssl x509 -noout -text | grep "Subject Alternative Name" -A 1
    X509v3 Subject Alternative Name:
      DNS:gardener-local-ha-multi-zone-control-plane

Real setup: both DNS and IP

  openssl s_client -connect 10.254.74.24:10250 -showcerts </dev/null 2>/dev/null | openssl x509 -noout -text | grep "Subject Alternative Name" -A 1
    X509v3 Subject Alternative Name:
      DNS:ip-10-254-74-24.eu-west-1.compute.internal, IP Address:10.254.74.24

However, federation still fails in the local setup, because the issuer of the
certificate is not trusted by Prometheus in the current configuration.