Make sure all seed alerts are not muted on the weekends
`mute_on_weekend` should be used for shoot alerts that are
customer issues and for which the Gardener team can't really
help. However, the Gardener team has full control of the
seeds so alerts on those should always be reacted upon: if
seed are having issues, then shoots can't be correctly managed.