[GEP-19] Migrate shoot Prometheus deployment (#9695)
* Integrate Prometheus deployment into `Shoot` controller

* Move Prometheus logging config into `prometheus` pkg

* Adapt components talking to shoot Prometheis

* Integrate `blackbox-exporter` deployment into `Shoot` controller

* Migrate central scrape configs

* Migrate central rules

* Remote write configuration

Note: Support for providing Queue configuration via gardenlet config has been dropped because it was of type `*string` which is no longer working now with Golang structs. We would have to reference the `monitoringv1.QueueConfig` type in the gardenlet config, however, I'd like to prevent this for now. It's anyways unclear whether this is used at all, and if, we can discuss how to cleanly allow modifying the configuration here. A better approach (now that we have a `Prometheus` CRD) could be that an extension mutates it and injects whatever configuration it wants.
In that sense, we might want to entirely drop support for providing remote write config via gardenlet config, and move everything out to extension responsibilities.

* Additional alertmanager configs

* Move target RBAC resources to dedicated `ManagedResource` managed in `prometheus` component

* Mount some secrets to well-known paths for backwards-compatibility with extensions

ETCD is mounted to prevent refactoring the monitoring config in this PR. Will follow in subsequent work/PR.

* Pass scrape configs and rules from components and extensions

We will refactor this to a clean usage of the Prometheus CRDs only in subsequent work/not in this PR.

* Delete no longer needed code

* Remove `StatefulSet` check from shoot care controller

The check is now performed by checking the `ManagedResource`

* Return copy of assets to prevent mutations of global variables

See test failures like this: https://prow.gardener.cloud/view/gs/gardener-prow/pr-logs/pull/gardener_gardener/9695/pull-gardener-e2e-kind/1785896315436142592

```
  {"level":"info","ts":"2024-05-02T05:20:58.970Z","logger":"shoot-test.test","msg":"Shoot not yet created","shoot":{"name":"e2e-managedseed","namespace":"garden"},"reason":"condition type ObservabilityComponentsHealthy is not true yet, had message error during apply of object \"monitoring.coreos.com/v1/PrometheusRule/shoot--local--e2e-hib/shoot-kube-kubelet\": namespaces \"shoot--local--e2e-hib\" not found with reason ApplyFailed"}

```

--> The shoot is called `e2e-managedseed`, but the `PrometheusRule` is targeting the `shoot--local--e2e-hib` namespace.

* Address PR review feedback