Use `last_over_time` instead of `max_over_time`
The usage of a `<aggregate>_over_time` function is only required to
check if there is data within the queried time window that triggers
the alert. If the query return data, then the value is always true.
Otherwise, the query returns nothing.

Therefore, it is meaningless to calculate the maximum over a set
of "trues". Instead, using `last_over_time` is a better approach
to check if the query returned data.