label workers with their expected operating system config key
Nodes with an outdated operating system config key are skipped while
checking whether a shoot is healthy. Currently, this requires botanist
to know how to calculate the OSCkey based on a node object, which leaks
implementation details. Instead, each node is now just labeled with the
expected config key. This allows for future extensions of the OSCkey
calculation without having to introduce additional labels on the node
objects.

By including the keyName in the operatingsystemconfig.Data struct, no
other component has to know how the keyName is calculated.