Add additional alert_relabel_configs to the garden Prometheus
The goal is to generate a new label for the shoot alerts containing such
shoot's dashboard URL so that we have a direct link to it in the
VictorOps alerts.

The shoot topology is used to discard alerts that do not correspond to
shoots. This is however undone in the next commit in favour of a
different strategy.