Fix shoot control plane federation regression (2) (#9769)
We missed to restrict the prometheusrules discovery to the
namespace of prometheus in #9764, which fixes a regression of #9695.

Without this fix, all the shoot control plane prometheus instances
process all the recording rules of all the other Prometheus instances,
such that the same (equivalent) recording rule is added up to 200 times.

Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>