Generalize shoot_dashboard_url label for all alerts with name and project
We can add the new shoot_dashboard_url label to all alerts that have
a name and a project because they probably reference a shoot, and we can
build the dashboard URL for them instead of specifically choosing alerts
with "shoot" topology. We only need to be aware that the dashboard URL
for shoots in the garden project (i.e., seeds that are shoots) has
subpath "/namespace/garden" and not "/namespace/garden-garden". To
circumvent this, a new alert relabel config is added, which overwrites
the previous one in case of the garden project and builds the
shoot_dashboard_url label correctly.