Remove resource limits for the observability components
Some time ago we decided to avoid using resource limits but somehow it was not
consistently implemented for these components.

The trigger for this change is a non self healing condition that we observed
recently: when the vertical pod autoscaling has an anomaly (not understood) and
it scales the resource requests of a pod to the lower bound recommendation
instead of the target recommendation (e.g. 1m core and 2.6MB for the blackbox
exporter), the limits are scaled proportionally (e.g. 13MB).

The problem is that such low memory limits do not manifest in OOM failures, but
the container itself can not be started in the first place because the `runc`
process is killed due to the out of memory condition on the cgroup. This
manifests in a RunContainerError in the pod status and can not self heal.

    State:          Waiting
      Reason:       RunContainerError
    Last State:     Terminated
      Reason:       StartError
      Message:      failed to create containerd task:
                    failed to create shim task: context canceled: unknown

With this change the limits are removed. Even if there is a momentary anomaly in
the vertical pod autoscaling, the pod will be able to start and the autoscaler
will correct the resource requests in the next iteration.