Address review comments: Skip TLS verification only for managed seeds
Use the garden client and try to look for a managed seed with the same name as
the seed. If it exists, it is a managed seed, otherwise it isn't.

The isManagedSeed flag is passed down to the AdditionalScrapeConfigs function
and golang text templating is used to set the insecure_skip_verify flag.

To test this in the KinD based local setup: instead of using the default local
project, create the shoot in the garden project (namespace) and upgrade it to a
managed seed with:

    apiVersion: seedmanagement.gardener.cloud/v1alpha1
    kind: ManagedSeed
    metadata:
      name: local
      namespace: garden
    spec:
      shoot:
        name: local
      gardenlet:
        bootstrap: BootstrapToken
        mergeWithParent: true
        config:
          seedConfig:
            spec:
              ingress:
                controller:
                  kind: nginx

Then check the scrape config both in the KinD cluster and in the managed seed:

    k get secrets prometheus-cache-additional-scrape-configs -o json | jq -r '
      .data."prometheus.yaml" | <at>base64d
    ' | grep skip

TLS verification is enabled in the KinD cluster but disabled in the managed
seed.