Introduce `shoot_dashboard_url` label (#9818)
* Add additional alert_relabel_configs to the garden Prometheus

The goal is to generate a new label for the shoot alerts containing such
shoot's dashboard URL so that we have a direct link to it in the
VictorOps alerts.

The shoot topology is used to discard alerts that do not correspond to
shoots. This is however undone in the next commit in favour of a
different strategy.

* Generalize shoot_dashboard_url label for all alerts with name and project

We can add the new shoot_dashboard_url label to all alerts that have
a name and a project because they probably reference a shoot, and we can
build the dashboard URL for them instead of specifically choosing alerts
with "shoot" topology. We only need to be aware that the dashboard URL
for shoots in the garden project (i.e., seeds that are shoots) has
subpath "/namespace/garden" and not "/namespace/garden-garden". To
circumvent this, a new alert relabel config is added, which overwrites
the previous one in case of the garden project and builds the
shoot_dashboard_url label correctly.

* Add unit tests for additional alert_relabel_configs

* [review] Address first set of comments

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* [review] Rename to AdditionalAlertRelabelConfigs

There was a typo in previous name.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>

* [review] Address second set of comments

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

---------

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>