Drop `Monitoring` struct from `shoot/types.go`
This seperation is not clear/does not bring benefits since the
`SystemComponents` struct also contains monitoring-related deployers.
To correct these, we would have to introduce a struct like
`MonitoringSystemComponents` or whatever, but I don't think that's
reasonable. Using the `ControlPlane` and `SystemComponents` alone seems
good enough and is also clear because it directly shows where the
respective component is deployed.