Label nodes with expected OperatingSystemConfig name (#9757)
* operatingsystemconfig: secret name is never nil

The only place that creates operatingsystemconfig.Data objects always
sets SecretName to a pointer. Thus, convert the string pointer to a
plain string.

* label workers with their expected operating system config key

Nodes with an outdated operating system config key are skipped while
checking whether a shoot is healthy. Currently, this requires botanist
to know how to calculate the OSCkey based on a node object, which leaks
implementation details. Instead, each node is now just labeled with the
expected config key. This allows for future extensions of the OSCkey
calculation without having to introduce additional labels on the node
objects.

By including the keyName in the operatingsystemconfig.Data struct, no
other component has to know how the keyName is calculated.

* worker: enforce the OSC exists during deploy

* test health check fallback for nodes without OSC key

* Revert "operatingsystemconfig: secret name is never nil"

This reverts commit bf5342c67c98938d32cb130fae4013adfe10a242.

* address review comments