The Gardener Dashboard is deployed in the runtime-garden cluster
Use the ingress domain of the runtime-garden cluster to reach
the Gardener Dashboard:

In the local setup, instead of

  https://dashboard.virtual-garden.local.gardener.cloud/healthz

the URL

  https://dashboard.ingress.runtime-garden.local.gardener.cloud/healthz

should be used to access the healthz endpoint of the Gardener Dashboard
via Istio + Nginx (TLS).

Note that in the dev landscape, the domains of the runtime and virtual
garden clusters are overlapping, so that this issue did not cause problems.

In the local setup, the domains of the runtime and virtual cluster are
different, so that it is important to pick the correct domain when
probing the Gardener Dashboard.