Use one single alert for all conditions: Remove previous alerts
This commit and the next discard most of previous work. The fundamental
change is that all seed condition are grouped into one alert only,
instead of having a specific alert for condition.

For the sake of change diff, this change is split in two commits: this
one removes previous alerts, and the next adds the new single alert.