Use one single alert for all conditions: Add new alert
This and previous commit discard most of previous work. The fundamental
change is that all seed condition are now grouped into one alert only,
instead of having a specific alert for condition.

For the sake of change diff, this change is split in two commits: the
previous one removed previous alerts, and this adds the new single alert.