Use the node IP instead
The previous approach is conceptually fine, but it would require an
additional network policy: 'allow-to-public-networks-local-setup'
(see previous commit), to allow the blackbox exporter to connect
to these endpoints.

    https://dashboard.ingress.runtime-garden.local.gardener.cloud/healthz
    https://gardener.virtual-garden.local.gardener.cloud/healthz

It turns out that we can also use the node IPs: the packets are routed
to the istio service and a new network policy is not needed.

Note that I'm not sure which component is mapping the port :443 to the
node port or service port.

Connecting to the <node IP>:<node port>

  openssl s_client -servername dashboard.ingress.runtime-garden.local.gardener.cloud -connect 172.18.0.4:30443 </dev/null 2>/dev/null -showcerts | openssl x509 -text | grep "X509v3 Subject Alternative Name" -A 1 | sed 's/,/\n/g'
    X509v3 Subject Alternative Name:
      DNS:dashboard.ingress.runtime-garden.local.gardener.cloud

works as expected.

Somehow connecting to <node IP>:<service port> also happens to work

  openssl s_client -servername dashboard.ingress.runtime-garden.local.gardener.cloud -connect 172.18.0.4:443 </dev/null 2>/dev/null -showcerts | openssl x509 -text | grep "X509v3 Subject Alternative Name" -A 1 | sed 's/,/\n/g'
    X509v3 Subject Alternative Name:
      DNS:dashboard.ingress.runtime-garden.local.gardener.cloud

So this approach is superior to the previous commit:

- does not require an additional network policy

  this network path is covered by the rule
  to-all-istio-ingresses-istio-ingressgateway-tcp-9443
  because it seems the "firewall" iptables rules are applied after
  the service port :443 is mapped to the pod port :9443

- fits well into the existing host mapping example for:

  garden.local.gardener.cloud