Rework seed alerts to consider all possible conditions (#9750)
* Revisit topology for current alerts

- Topology `shoot` to refer a seed is incorrect.
- Change topology from `garden` to `seed` for
  those alerts checking seed conditions.

* Remove unnecessary labels `is_compliant` and `has_user_errors`

We do not want to suppress alerts on seeds that are managed by the
Gardener team, even if the seeds are not compliant or the errors are
user errors.

* Make sure `max_over_time` is used in the alerts in case of flapping

Without it, flapping conditions would probably go undetected because
the alerts would be reset each time.

For example, when the seed is reconciling its artifacts, some of its
managed resources can enter the progressing state for a brief period,
and to avoid alerts from being restarted, the `max_over_time` function
is now introduced to smooth out this behaviour. This is combined with
the subquery syntax [1] to query over the last minutes only.

This change is also reducing the time before firing an alert in the `for`
entry from 30m to 10m for some alerts to alert more aggressively if any of
those conditions fail. An exception is the alert `SeedAPIServerUnavailable`,
which was already firing after 2m, but now changed to 3m to preserve the
same behaviour after the change introducing `max_over_time` to smooth
out possible flappings.

[1] https://prometheus.io/docs/prometheus/latest/querying/basics/#subquery

* Use `last_over_time` instead of `max_over_time`

The usage of a `<aggregate>_over_time` function is only required to
check if there is data within the queried time window that triggers
the alert. If the query return data, then the value is always true.
Otherwise, the query returns nothing.

Therefore, it is meaningless to calculate the maximum over a set
of "trues". Instead, using `last_over_time` is a better approach
to check if the query returned data.

* Make sure all seed alerts are not muted on the weekends

`mute_on_weekend` should be used for shoot alerts that are
customer issues and for which the Gardener team can't really
help. However, the Gardener team has full control of the
seeds so alerts on those should always be reacted upon: if
seed are having issues, then shoots can't be correctly managed.

* Fix description for the APIServerAvailable alert

Add the list of active alerts as in other alerts by `garden_shoot_condition`

* Add new alerts to cover all possible seed conditions

* Homogenize summary and description for seed condition alerts

- Use common text for all alerts regarding seed conditions
- Fix description for the APIServerAvailable alert by adding
  the list of active alerts as in other alerts by `garden_shoot_condition`

* Add promtool tests for new seed condition alerts

The different condition states are mapped to numbers by the
`gardener-metrics-exporter`:

```
2: Progressing
1: True
0: False
-1: Unknown
```

* Use one single alert for all conditions: Remove previous alerts

This commit and the next discard most of previous work. The fundamental
change is that all seed condition are grouped into one alert only,
instead of having a specific alert for condition.

For the sake of change diff, this change is split in two commits: this
one removes previous alerts, and the next adds the new single alert.

* Use one single alert for all conditions: Add new alert

This and previous commit discard most of previous work. The fundamental
change is that all seed condition are now grouped into one alert only,
instead of having a specific alert for condition.

For the sake of change diff, this change is split in two commits: the
previous one removed previous alerts, and this adds the new single alert.

* Add unit test for multiple shoots and alerts

This unit test shows that, if two different seeds, have failing
conditions then two different alerts pop up as well.

* Mute new alert on seed conditions

Mute this alert on weekends as we gain experience on how noisy
it will be on live and canary. The goal, though, is that we
eventually unmute it.

* Alert after 10 minutes instead of 6 minutes

After testing this change on dev, we found alerting becomes
a bit too noisy so we choose to increase the for clause up
to 10 minutes again.

* Remove empty line at the end of the alert description

For better alert formatting, alert descriptions are not supposed to
contain an empty line at the very end.

* [review] Address review comments

- Unroll the time series in the tests so it's simpler.
- Add comment on how conditions states are mapped into numbers.