Adjust the blackbox exporter config for the Gardener Dashboard
If the Dashboard Certificate is issued by Gardener,
the 'http_gardener_dashboard' blackbox exporter module
needs to specify the Gardener CA as certificate authority.

In the local setup, the certificate of the Dashboard is issued by Gardener
and without this change the probes fail.

In a real setup, we use a Let's Encrypt issued certificate for the ingress
of the Gardener Dashboard. This can be verified by the Root CA's that
are available in the blackbox exporter docker image under
'/etc/ssl/certs/ca-certificates.crt', contributed by the second layer of
the busybox image.

So this change helps to avoid failures in the local setup and is a no-op
for real setups that do not use Gardener issued certificates for the
Gardener Dashboard.