Fix: Use garden cluster API reader when deleting the bootstrap token
Earlier, it was tried to use the client, however, the client (and its
backing cache) is not yet started at this point (the `gardenCluster` is
only added to the manager (thus, started) a few lines later at L327).

Hence, we should use the API reader (not backed by cache that must be
started) when calling this function.