Fix the {virtual|runtime}-garden probes in the local gardener-operator setup (#9832)
* The Gardener Dashboard is deployed in the runtime-garden cluster

Use the ingress domain of the runtime-garden cluster to reach
the Gardener Dashboard:

In the local setup, instead of

  https://dashboard.virtual-garden.local.gardener.cloud/healthz

the URL

  https://dashboard.ingress.runtime-garden.local.gardener.cloud/healthz

should be used to access the healthz endpoint of the Gardener Dashboard
via Istio + Nginx (TLS).

Note that in the dev landscape, the domains of the runtime and virtual
garden clusters are overlapping, so that this issue did not cause problems.

In the local setup, the domains of the runtime and virtual cluster are
different, so that it is important to pick the correct domain when
probing the Gardener Dashboard.

* WIP: Add a core DNS entry for the {virtual|runtime}-garden domains

In the local gardener operator setup, the DashboardDown and ApiServerDown alerts
are firing. These are false positives.

The Garden Prometheus probes the Gardener API server and Gardener Dashboard via
the blackbox-exporter using these URLs:

    https://dashboard.ingress.runtime-garden.local.gardener.cloud/healthz
    https://gardener.virtual-garden.local.gardener.cloud/healthz

These domains can not be resolved by the core DNS server of the KinD cluster, so
the probes fail.

To fix this, we need to add a core DNS entry for these domains.

In a real setup, these domains are resolved to the external IP of the load
balancer service of the istio ingress gateway.

In the KinD cluster, that corresponds to the IP of the docker0 device. Using the
IP of the docker0 device as the target IP for the core DNS entries, the network
path of the packets touches all the relevant components of the Gardener setup:
docker port mapping, node port mapping to the istio ingress gateway service, and
finally the Gardener API server or the dashboard.

```
ip addr show docker0

4: docker0: <NO-CARRIER,BROADCAST,MULTICAST,UP> mtu 1500 qdisc noqueue state DOWN
    link/ether 02:42:c1:4a:50:a4 brd ff:ff:ff:ff:ff:ff
    inet 172.17.0.1/16 brd 172.17.255.255 scope global docker0
      valid_lft forever preferred_lft forever
    inet6 fe80::42:c1ff:fe4a:50a4/64 scope link
      valid_lft forever preferred_lft forever
```

```
ip addr show docker0 \
| awk '/inet .* docker0/{print substr($2, 0, index($2, "/") - 1)}'

172.17.0.1
```

The docker port mapping:

```
docker inspect gardener-operator-local-control-plane | jq -r '
  .[].NetworkSettings.Ports."30443/tcp".[].HostPort'

443 -> 30443
```

The istio ingress gateway service node port mapping:

```
k -n virtual-garden-istio-ingress get svc istio-ingressgateway -o json \
| jq '.spec.ports[]'

{
  "name": "tcp",
  "nodePort": 30443,
  "port": 443,
  "protocol": "TCP",
  "targetPort": 9443
}
```

and finally the Gardener API server or the Gardener dashboard:

```
k get vs (| format...)

NAME:     nginx-ingress-controller-0
GATEWAYS: ["nginx-ingress-controller"]
HOSTS:    ["*.ingress.runtime-garden.local.gardener.cloud"]

NAME:     virtual-garden-kube-apiserver
GATEWAYS: ["virtual-garden-kube-apiserver"]
HOSTS:    ["api.virtual-garden.local.gardener.cloud","gardener.virtual-garden.local.gardener.cloud"]
```

---

The following network policy would need to be deployed in the local setup to
allow the blackbox exporter to access the conceptually public IPs that happen to
be internal IPs in the local setup:

```
apiVersion: networking.k8s.io/v1
kind: NetworkPolicy
metadata:
  name: allow-to-public-networks-local-setup
  namespace: garden
spec:
  egress:
  - to:
    - ipBlock:
        cidr: 172.17.0.0/16
  podSelector:
    matchLabels:
      networking.gardener.cloud/to-public-networks: allowed
  policyTypes:
  - Egress
```

* Use the node IP instead

The previous approach is conceptually fine, but it would require an
additional network policy: 'allow-to-public-networks-local-setup'
(see previous commit), to allow the blackbox exporter to connect
to these endpoints.

    https://dashboard.ingress.runtime-garden.local.gardener.cloud/healthz
    https://gardener.virtual-garden.local.gardener.cloud/healthz

It turns out that we can also use the node IPs: the packets are routed
to the istio service and a new network policy is not needed.

Note that I'm not sure which component is mapping the port :443 to the
node port or service port.

Connecting to the <node IP>:<node port>

  openssl s_client -servername dashboard.ingress.runtime-garden.local.gardener.cloud -connect 172.18.0.4:30443 </dev/null 2>/dev/null -showcerts | openssl x509 -text | grep "X509v3 Subject Alternative Name" -A 1 | sed 's/,/\n/g'
    X509v3 Subject Alternative Name:
      DNS:dashboard.ingress.runtime-garden.local.gardener.cloud

works as expected.

Somehow connecting to <node IP>:<service port> also happens to work

  openssl s_client -servername dashboard.ingress.runtime-garden.local.gardener.cloud -connect 172.18.0.4:443 </dev/null 2>/dev/null -showcerts | openssl x509 -text | grep "X509v3 Subject Alternative Name" -A 1 | sed 's/,/\n/g'
    X509v3 Subject Alternative Name:
      DNS:dashboard.ingress.runtime-garden.local.gardener.cloud

So this approach is superior to the previous commit:

- does not require an additional network policy

  this network path is covered by the rule
  to-all-istio-ingresses-istio-ingressgateway-tcp-9443
  because it seems the "firewall" iptables rules are applied after
  the service port :443 is mapped to the pod port :9443

- fits well into the existing host mapping example for:

  garden.local.gardener.cloud

* Adjust the blackbox exporter config for the Gardener Dashboard

If the Dashboard Certificate is issued by Gardener,
the 'http_gardener_dashboard' blackbox exporter module
needs to specify the Gardener CA as certificate authority.

In the local setup, the certificate of the Dashboard is issued by Gardener
and without this change the probes fail.

In a real setup, we use a Let's Encrypt issued certificate for the ingress
of the Gardener Dashboard. This can be verified by the Root CA's that
are available in the blackbox exporter docker image under
'/etc/ssl/certs/ca-certificates.crt', contributed by the second layer of
the busybox image.

So this change helps to avoid failures in the local setup and is a no-op
for real setups that do not use Gardener issued certificates for the
Gardener Dashboard.