`gardenlet` can create `Gardenlet` object during start-up
For convenience reasons, the `gardenlet` can create the `Gardenlet`
object in the garden cluster itself if the Helm chart values have been
configured with the OCI repository config.

This way, a human just needs to deploy the Helm chart and does not need
to explicitly create the `Gardenlet` object in the garden cluster. This
improves conveniences and the user experience.