`gardenlet` Self-Upgrades For Unmanaged `Seed`s (#9874)
* Nit: Fix some typos

* Fix: Use garden cluster API reader when deleting the bootstrap token

Earlier, it was tried to use the client, however, the client (and its
backing cache) is not yet started at this point (the `gardenCluster` is
only added to the manager (thus, started) a few lines later at L327).

Hence, we should use the API reader (not backed by cache that must be
started) when calling this function.

* Move `gardenlet` Helm chart tests to `./charts/gardener/gardenlet`

Having the test code next to the implmentation simplifies maintenance.

* Documentation

* Add new `seedmanagement.gardener.cloud/v1alpha1.Gardenlet` API

types only, rest is following in subsequent commits

* Defaulting and validation

* [make generate]

* Register `Gardenlet` API in `gardener-apiserver`

* Prevent creation of `Gardenlet` when `ManagedSeed` already exists (and vice versa)

* Push `gardenlet` Helm chart to local registry in Skaffold builds

* Adapt chart applier to support Apply/Delete from archive

Earlier, it only supported Apply/Delete from an embedded FS.

* Adapt seed authorizer plugin

* Adapt seed restriction admission webhook

This allows creation of bootstrap token secrets.

* Make chart value prep code in `ManagedSeed` controller reusable

The `Gardenlet` controller will use this since it is basically the same
logic that is performed.

* Add `Gardenlet` controller options to config API

* Controller business logic

* Integration test

* `gardenlet` can create `Gardenlet` object during start-up

For convenience reasons, the `gardenlet` can create the `Gardenlet`
object in the garden cluster itself if the Helm chart values have been
configured with the OCI repository config.

This way, a human just needs to deploy the Helm chart and does not need
to explicitly create the `Gardenlet` object in the garden cluster. This
improves conveniences and the user experience.

* Address PR review feedback

* Address PR review feedback

* Address PR review feedback

* Address PR review feedback